Remove dependency `quickwit-query` <- `quickwit-proto`
