+++ quickwit/quickwit-doc-mapper/Cargo.toml
@@ -20,8 +20,6 @@ itertools = { workspace = true }
 mockall = { workspace = true, optional = true }
 nom = { workspace = true }
 once_cell = { workspace = true }
-quickwit-datetime = { workspace = true }
-quickwit-query = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -32,11 +30,13 @@ tracing = { workspace = true }
 typetag = { workspace = true }
 utoipa = { workspace = true }
 
+quickwit-datetime = { workspace = true }
+quickwit-query = { workspace = true }
+
 [dev-dependencies]
 criterion = { workspace = true }
 matches = { workspace = true }
 proptest = { workspace = true }
-quickwit-proto = { workspace = true }
 serde_yaml = { workspace = true }
 time = { workspace = true }
