+++ quickwit/quickwit-indexing/src/actors/merge_executor.rs
@@ -694,7 +694,7 @@ mod tests {
                 index_uid: index_uid.to_string(),
                 start_timestamp: None,
                 end_timestamp: None,
-                query_ast: quickwit_proto::qast_helper(delete_query, &["body"]),
+                query_ast: quickwit_query::query_ast::qast_helper(delete_query, &["body"]),
             })
             .await?;
         let split_metadata = metastore
