+++ quickwit/quickwit-metastore/Cargo.toml
@@ -38,6 +38,7 @@ quickwit-common = { workspace = true }
 quickwit-config = { workspace = true }
 quickwit-doc-mapper = { workspace = true }
 quickwit-proto = { workspace = true }
+quickwit-query = { workspace = true }
 quickwit-storage = { workspace = true }
 
 [dev-dependencies]
