+++ quickwit/quickwit-metastore/src/metastore/file_backed_metastore/file_backed_index/mod.rs
@@ -88,7 +88,7 @@ impl quickwit_config::TestableForRegression for FileBackedIndex {
                 index_uid: "index:11111111111111111111111111".to_string(),
                 start_timestamp: None,
                 end_timestamp: None,
-                query_ast: quickwit_proto::qast_helper("Harry Potter", &["body"]),
+                query_ast: quickwit_query::query_ast::qast_helper("Harry Potter", &["body"]),
             }),
         };
         FileBackedIndex::new(index_metadata, splits, vec![delete_task])
