+++ quickwit/quickwit-proto/Cargo.toml
@@ -14,7 +14,6 @@ anyhow = { workspace = true }
 opentelemetry = { workspace = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
-quickwit-query = { workspace = true }
 tonic = { workspace = true }
 tracing = { workspace = true }
 tracing-opentelemetry = { workspace = true }
