+++ quickwit/quickwit-query/src/lib.rs
@@ -29,12 +29,11 @@
 // documentation.
 
 mod elastic_query_dsl;
+mod error;
 mod json_literal;
+mod not_nan_f32;
 pub mod query_ast;
 mod tokenizers;
-
-mod error;
-mod not_nan_f32;
 
 pub use elastic_query_dsl::{ElasticQueryDsl, OneFieldMap};
 pub use error::InvalidQuery;
