+++ quickwit/quickwit-serve/src/delete_task_api/handler.rs
@@ -23,8 +23,8 @@ use quickwit_config::build_doc_mapper;
 use quickwit_janitor::error::JanitorError;
 use quickwit_metastore::{Metastore, MetastoreError};
 use quickwit_proto::metastore::{DeleteQuery, DeleteTask};
-use quickwit_proto::{query_ast_from_user_text, IndexUid, SearchRequest};
-use quickwit_query::query_ast::QueryAst;
+use quickwit_proto::{IndexUid, SearchRequest};
+use quickwit_query::query_ast::{query_ast_from_user_text, QueryAst};
 use serde::Deserialize;
 use warp::{Filter, Rejection};
