+++ quickwit/quickwit-serve/src/search_api/rest_handler.rs
@@ -23,7 +23,8 @@ use std::sync::Arc;
 use futures::stream::StreamExt;
 use hyper::header::HeaderValue;
 use hyper::HeaderMap;
-use quickwit_proto::{query_ast_from_user_text, OutputFormat, ServiceError, SortOrder};
+use quickwit_proto::{OutputFormat, ServiceError, SortOrder};
+use quickwit_query::query_ast::query_ast_from_user_text;
 use quickwit_search::{SearchError, SearchResponseRest, SearchService};
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use serde_json::Value as JsonValue;
